+++ src/main/scala/grizzled/collection/MultiIterator.scala
@@ -3,9 +3,13 @@ package grizzled.collection
 /** An iterator that iterates, serially, over the contents of multiple other
   * iterators.
   *
+  * '''This class is no longer necessary.''' Just use the `Iterator` operator
+  * `++`, as outlined in
+  * [[http://stackoverflow.com/a/16315251/53495 this StackOverflow answer]].
+  *
   * @param iterators  the iterators to wrap
   */
-@SuppressWarnings(Array("org.brianmckenna.wartremover.warts.Var"))
+@deprecated("Use Iterator++. See http://stackoverflow.com/a/16315251/53495", "2.3.2")
 class MultiIterator[+T](iterators: Iterator[T]*) extends Iterator[T] {
   private[this] val iteratorList: List[Iterator[T]] = iterators.toList
   private[this] var current = iteratorList.head